requirements.txt
asgiref==3.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-filter==21.1
django-jazzmin==2.4.9
django-js-asset==1.2.2
django-richtextfield==1.6.1
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5
